requirements.txt
litellm>=0.1.45
python-dotenv>=1.0.0
